+++ src/Configuration.cpp
@@ -54,6 +54,7 @@ Configuration::Configuration(wxDC *dc, InitOpt options) : m_dc(dc) {
     {
       m_styleNames[TS_CODE_DEFAULT        ] = _("Code Default"); 
       m_styleNames[TS_VARIABLE            ] = _("Output: Variable names");
+      m_styleNames[TS_OPERATOR            ] = _("Output: Operators");
       m_styleNames[TS_NUMBER              ] = _("Output: Numbers and Digits");
       m_styleNames[TS_FUNCTION            ] = _("Output: Function names");
       m_styleNames[TS_SPECIAL_CONSTANT    ] = _("Output: Special constants (%e,%i)");
@@ -108,6 +109,7 @@ Configuration::Configuration(wxDC *dc, InitOpt options) : m_dc(dc) {
   if(m_2dMathStyles.empty())
     {
       m_2dMathStyles.push_back(TS_VARIABLE);
+      m_2dMathStyles.push_back(TS_OPERATOR);
       m_2dMathStyles.push_back(TS_NUMBER);
       m_2dMathStyles.push_back(TS_FUNCTION);
       m_2dMathStyles.push_back(TS_SPECIAL_CONSTANT);
@@ -1150,6 +1152,7 @@ void Configuration::ReadStyles(const wxString &file) {
   m_styles[TS_STRING].Read(config, "Style/String/");
   m_styles[TS_ASCIIMATHS].Read(config, "Style/ASCIImaths/");
   m_styles[TS_VARIABLE].Read(config, "Style/Variable/");
+  m_styles[TS_OPERATOR].Read(config, "Style/Operator/");
   m_styles[TS_FUNCTION].Read(config, "Style/Function/");
   m_styles[TS_HIGHLIGHT].Read(config, "Style/Highlight/");
   m_styles[TS_TEXT_BACKGROUND].Read(config, "Style/Background/");
@@ -1556,6 +1559,7 @@ void Configuration::WriteStyles(wxConfigBase *config) {
   m_styles[TS_STRING].Write(config, "Style/String/");
   m_styles[TS_ASCIIMATHS].Write(config, "Style/ASCIImaths/");
   m_styles[TS_VARIABLE].Write(config, "Style/Variable/");
+  m_styles[TS_OPERATOR].Write(config, "Style/Operator/");
   m_styles[TS_FUNCTION].Write(config, "Style/Function/");
   m_styles[TS_HIGHLIGHT].Write(config, "Style/Highlight/");
   m_styles[TS_TEXT_BACKGROUND].Write(config, "Style/Background/");