+++ src/MathParser.cpp
@@ -170,6 +170,18 @@ MathParser::MathParser(Configuration *cfg, const wxString &zipfile) {
 
 MathParser::~MathParser() {}
 
+std::unique_ptr<Cell> MathParser::ParseVariableNameTag(wxXmlNode *node){
+  auto children = node->GetChildren();
+  if (children != NULL)
+    {
+      if(m_configuration->IsOperator(node->GetContent()))
+	return ParseText(children, TS_OPERATOR);
+      else
+	return ParseText(children, TS_VARIABLE);
+    }
+  return ParseText(children, TS_VARIABLE);
+}
+
 std::unique_ptr<Cell> MathParser::ParseHiddenOperatorTag(wxXmlNode *node) {
   auto retval = ParseText(node->GetChildren());
   retval->SetHidableMultSign(true);