+++ NEWS.md
@@ -5,6 +5,7 @@
  * Dropped the dependency on an [internal] ww898
  * More consistent text style handling
  * A try to improve locale management on the Mac platform
+ * A separate style for operators (#1002)
 
 # 22.12.0
  * A cleaner status bar with double-click actions